+++ src/stocks/parse.ts
@@ -1,21 +1,8 @@
 import { createReadStream, PathLike } from "fs"
 import { parse } from "csv-parse"
 
-import { Type, Timestamp, Transaction } from "../types"
+import { Transaction } from "./types"
 import { getISODate, getTimestampByDate, parseNumber } from "../utils"
-
-const SELL_FEE_1 = 0.01
-const SELL_FEE_2 = 0.02
-
-function getSellFee(timestamp: Timestamp) {
-    const newFeeDate = getTimestampByDate("2022-05-27")
-
-    if (timestamp < newFeeDate) {
-        return SELL_FEE_1
-    }
-
-    return SELL_FEE_2
-}
 
 export async function parseStatement(filePath: PathLike) {
     const promise = (): Promise<Transaction[]> =>
@@ -26,12 +13,6 @@ export async function parseStatement(filePath: PathLike) {
                 .pipe(parse({ delimiter: ",", from_line: 2 }))
                 .on("data", (csvRow) => {
                     const timestamp = getTimestampByDate(csvRow[0])
-                    const type = csvRow[2]
-
-                    const sellFee =
-                        type === Type.Sell ? getSellFee(timestamp) : null
-
-                    const fee = csvRow[8] ? parseNumber(csvRow[8]) : sellFee
 
                     const transaction: Transaction = {
                         date: getISODate(csvRow[0]),
@@ -45,7 +26,6 @@ export async function parseStatement(filePath: PathLike) {
                         amount: parseNumber(csvRow[5]),
                         currency: csvRow[6],
                         fxRate: parseNumber(csvRow[7]),
-                        fee,
                     }
 
                     statement.push(transaction)