+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Fix deal types
 -   Fix date parsing
 -   Fix fee handling
+-   Fix fee calculation
 
 ## [v0.7.0] - 2022-04-18